+++ docs/CHANGELOG.md
@@ -31,4 +31,5 @@
 - Konfigurace připojení k databázi v souboru application.properties
 - Úspěšné spuštění Spring Boot aplikace a ověření připojení k databázi aegiscore_banking
 - Vytvoření package pro entity a implementace entity třídy Client -> aktualizace dokumentace (PROJECT_CONTEXT.md)
+- Úprava entity třídy Client zahrnutím @GeneratedValue(strategy = GenerationType.IDENTITY) a vytovření první repository vrstvy pro entitu Client -> aktualizace dokumentace (PROJECT_CONTEXT.md)
 -